+++ app/assets/stylesheets/profiles.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the profiles controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: https://sass-lang.com/
