+++ app/views/authentication/_form.html.slim
@@ -1,4 +1,4 @@
-= form_with( url: login_path, model: user, local: true, method: :post ) do |form|
+= form_with( url: login_path, local: true, method: :post ) do |form|
   .form-group
     = form.label :email
     = form.text_field :email, class: "form-control", type: "email", id: "email", autofocus: true
