+++ app/views/authentication/new.html.slim
@@ -4,4 +4,4 @@ p Find me in app/views/authentication/new.html.slim
 .container
   .row
     .col-md-10.offset-md-1.col-lg-8.offset-lg-2
-      = render 'form', user: @user
+      = render 'form'
