+++ app/views/profiles/edit.html.slim
@@ -3,7 +3,7 @@ p Find me in app/views/users/create.html.slim
 
 div.row
   div.col-md-10.offset-md-1.col-lg-8.offset-lg-2
-    = form_with(model: @user, local: true) do |form|
+    = form_with(model: @user, url: profile_path, local: true) do |form|
       = render 'shared/error_messages', model: form.object
       .form-group
         = form.label :name
